requirements.txt
Django==3.2.19
djangorestframework==3.15.2
django-filter==22.1
psycopg2-binary==2.9.6
python-decouple==3.8